workers/layout_worker_marker/requirements.txt
marker-pdf>=1.10.2
torch>=2.4.0
